Refactor registration page
