+++ source/components/FormSideBar/styles.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/variables.scss';
+@import '../../styles/variables.scss';
 
 $link-height: 20px;
 $link-yellow: #fcca88;
