+++ source/components/MultiPageForm/styles.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/variables.scss';
+@import '../../styles/variables.scss';
 
 $form-width: 500px;
