+++ source/scenes/Registration/Panes/Submit.jsx
@@ -7,13 +7,11 @@ import loader from 'assets/loader.svg';
 import uploadResume from 'services/api/upload';
 import register from 'services/api/registration';
 import { invalidateData } from 'services/registration/actions';
-import FormTransition from '../FormTransition';
-import FormContext from '../../../FormContext';
+import FormContext from '../FormContext';
 
 import './SubmitStyles.scss';
 
 type Props = {
-  visible: boolean,
   resumeDirty: boolean,
   dataDirty: boolean,
   reset: () => void,
@@ -83,7 +81,6 @@ class StudentInfo extends Component<Props, State> {
   }
 
   render() {
-    const { visible } = this.props;
     const { redirect, fetching, error, success } = this.state;
     let body = (
       <FormContext.Consumer>
@@ -119,11 +116,7 @@ class StudentInfo extends Component<Props, State> {
       );
     }
 
-    return (
-      <FormTransition visible={visible} uid="submission-page">
-        <div className="submission-page">{body}</div>
-      </FormTransition>
-    );
+    return <div className="submission-page">{body}</div>;
   }
 }
