+++ source/scenes/Registration/check.js
